Temp fix bug in errorList components
The ErrorList and InlineError components display errors
from unsuccessful requests that the Form object sends to
the server. The Form object must follow an interface which
is not yet provided by FormBuilder. Thus, create a temp
workaround to avoid errors caused by a Form object that
does not follow the proper interface.