change how errors are displayed
FormBuilder was getting the form errors from methods
of the 'model' object that is passed as a prop. The goal
of that was to make it easier to display errors when using
an object makes requests and handles errors by itself. So,
since the object would already sync with the user input, it
was convenient to use the same object (since it had access
to te data) to have methods to make API calls.

However, that was a bad thing. It required the model object
to have specific methods for the errors to be shown. This
make things less flexible and hard to extend. Therefore, I
decided to just add two new props: errors, and errorMessage.
Those props directly pass the errors to formBuilder, so there
are no requirements for the model prop.

From now, I'll try to make things more simple, flexible, and
with little requirements.