adapt tests to the v-model feature, which causes new behavior
Tests need to be adapted because the behavior of VueFormBuilder now  has
changed. For example, if `defaults` changes, it  no  longers  trigger  a
change in the current field values. Only if `modelValue` changes that it
does so. Moreover, if `modelValue` is set, it overwrites the  `defaults`
as the initial values for the fields.